jxl-render: Refactor
